+++ src/com/nadajp/littletalkers/MainActivity.java
@@ -19,7 +19,7 @@ public class MainActivity extends Activity implements AddKidListener
       setContentView(R.layout.activity_main);
 
       // Find out from shared preferences whether there are any kids yet
-      long kidId = Prefs.getKidId(this, -1);
+      int kidId = Prefs.getKidId(this, -1);
       //Log.i(DEBUG_TAG, "Kid Id in Main: " + kidId);
 
       /* If no kids have been added yet, go to AddKidActivity
